+++ installer/graphics.c
@@ -1,5 +1,6 @@
 #include <kernel.h>
 #include <stdlib.h>
+#include <malloc.h>
 #include <stdio.h>
 #include <string.h>
 #include <wchar.h>